Extract conversion test to separate file.
